requirements.txt
requests
python-dotenv
typer
pandas
